refactor(legion): extract shared toNum/get into format.ts
snapshot.ts and providers.ts each defined identical toNum/get helpers; hoist them
to the shared format module to prevent drift as provider types expand.
